refactor: keep credential modules within size policy
